factoring out some key report generation functions
